feat(new-ui): export our own ThemeProvider
so there is always a default theme, consumer doesn't need to even have SC as dependency
