--- packages/new-polymath-ui/src/styles/index.ts
@@ -1,23 +1,0 @@
-import * as styledComponents from 'styled-components';
-import { ThemedStyledComponentsModule } from 'styled-components';
-
-import { ThemeInterface } from './theme';
-
-const {
-  default: styled,
-  css,
-  createGlobalStyle,
-  withTheme,
-  keyframes,
-  ThemeProvider,
-} = styledComponents as ThemedStyledComponentsModule<ThemeInterface>;
-
-export {
-  css,
-  createGlobalStyle,
-  keyframes,
-  ThemeProvider,
-  withTheme,
-  ThemeInterface,
-};
-export default styled;
