+++ packages/new-polymath-ui/src/styles/theme.ts
@@ -1,30 +1,6 @@
 import { types } from '@polymathnetwork/new-shared';
 import { transparentize, darken } from 'polished';
-
-interface Scale<TValue> {
-  [id: string]: TValue;
-}
-
-export interface ThemeInterface {
-  breakpoints: [string];
-  fontSizes: Scale<string>;
-  fontFamilies: any;
-  lineHeights: any;
-  fontWeights: any;
-  space: Scale<string>;
-  zIndexes: any;
-  colors: any;
-  headings: any;
-  buttons: any;
-  transitions: any;
-  maxWidth: string;
-  inputs: any;
-  links: any;
-  header: any;
-  footer: any;
-  sidebar: any;
-  tokens: any;
-}
+import { Scale } from './types';
 
 export const breakpoints = ['42.5em', '64em', '80em']; // 680px, 1024px, 1280px
